Starting Work after FYP-1 (Presentation 2)

+++ database/migrations/2019_08_18_000002_create_states_table.php
@@ -14,9 +14,9 @@ class CreateStatesTable extends Migration
     public function up()
     {
         Schema::create('states', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
-            $table->unsignedInteger('country_id')->index();
+            $table->unsignedBigInteger('country_id')->index();
             $table->timestamps();
 
             $table->foreign('country_id')->references('id')->on('countries')->onDelete('cascade');
